Adding link checker on commit
Gitbook works by creating a "change request" in the gitbook frontend. In the back, when the change request is merged it seems to be just a simple commit to github. 

I'm adding a link checker that is triggered on commit. The `fail` flag is set to false as there are many broken links due to content pointing to Kovan. I don't want to delete the link or point to goerli cuz all the content is still talking about kovan. Also there are broken images which if I delete I lose the section that require an image (which I want to fix at some point). so for now, I just want to see that there are no new broken links, but allow gitbook to commit without any problem.